+++ app/sprinkles/account/locale/zh_CN/validate.php
@@ -1,0 +1,19 @@
+<?php
+/**
+ * UserFrosting (http://www.userfrosting.com)
+ *
+ * @link      https://github.com/userfrosting/UserFrosting
+ * @license   https://github.com/userfrosting/UserFrosting/blob/master/licenses/UserFrosting.md (MIT License)
+ *
+ * Chinese message token translations for the 'account' sprinkle.
+ *
+ * @package userfrosting\i18n\zh_CN
+ * @author @BruceGui (https://github.com/BruceGui)
+ */
+
+return [
+	"VALIDATE" => [
+        "PASSWORD_MISMATCH" => "密码不一致.",
+        "USERNAME" => "用户名必须以小写字母, 数字, '.', '-', 和 '_'组成."
+    ]
+];