+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Support for `withPivot` on `Unique` relationships (in tertiary models)
 - Factor out common code from `PermissionUserSprunje` into `UserSprunje`
 - Table of user permissions on user info page
+- Add Chinese translations
 
 ## v4.1.6-alpha
 - Fix missing permission check when `uri_account_settings` is not in role (#768)